refactor(benchmark): replace add_prompt with a single add_prompts method
add_prompt was a singular special-case of adding prompts. Consolidate it
and the private _add_prompts into one public, index-matched add_prompts
(plural) that validates, uploads concurrently, and handles partial
failures. create_new_benchmark now just creates the benchmark and
delegates, so prompt validation lives in one place. Docs updated.
